Addressing review comments - round 2
