+++ core/src/main/kotlin/net/corda/core/internal/X500Utils.kt
@@ -1,4 +1,3 @@
-@file:JvmName("CordaX500NameUtils")
 @file:KeepForDJVM
 
 package net.corda.core.internal
@@ -65,7 +64,7 @@ private val X500_NAME_PARTS = listOf(
         BCStyle.NAME)
 
 /**
- * Translates the X500Principal instance to the X500Name object. Additionally, it sorts the X500 name parts in the order
+ * Converts the X500Principal instance to the X500Name object. Additionally, it sorts the X500 name parts in the order
  * imposed by the X500_NAME_PARTS list.
  */
 fun X500Principal.toOrderedX500Name(supportedAttributes: Set<ASN1ObjectIdentifier> = emptySet()): X500Name {
@@ -73,9 +72,9 @@ fun X500Principal.toOrderedX500Name(supportedAttributes: Set<ASN1ObjectIdentifie
 }
 
 /**
- * Translates the X500Principal instance to the X500Name object.
+ * Converts the X500Principal instance to the X500Name object.
  */
-fun X500Principal.toX500Name() = X500Name.getInstance(this.encoded)
+fun X500Principal.toX500Name(): X500Name = X500Name.getInstance(this.encoded)
 
 private fun buildOrderedX500Name(attributes: Map<ASN1ObjectIdentifier, ASN1Encodable?>): X500Name {
     return X500NameBuilder(BCStyle.INSTANCE).apply {
@@ -98,7 +97,7 @@ private fun buildOrderedX500Name(attributes: Map<ASN1ObjectIdentifier, ASN1Encod
  *
  */
 fun X500Principal.toAttributesMap(supportedAttributes: Set<ASN1ObjectIdentifier> = emptySet()): Map<ASN1ObjectIdentifier, ASN1Encodable> {
-    val x500Name = X500Name.getInstance(this.encoded)
+    val x500Name = this.toX500Name()
     val attrsMap: Map<ASN1ObjectIdentifier, ASN1Encodable> = x500Name.rdNs
             .flatMap { it.typesAndValues.asList() }
             .groupBy(AttributeTypeAndValue::getType, AttributeTypeAndValue::getValue)
